ml/pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "ml" },
]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.6.0.66"
hydra-core = "^1.2.0"
tqdm = "^4.64.1"
ffmpeg-python = "^0.2.0"
python-speech-features = "^0.6"
pydub = "^0.25.1"
scipy = "^1.9.3"
torch = "^1.13.0"
torchvision = "^0.14.0"
torchaudio = "^0.13.0"
pytorch-lightning = "^1.8.3.post1"
pillow = "^9.3.0"
torchmetrics = "^0.11.0"
albumentations = "^1.3.0"
timm = "^0.6.12"
pytorch-pretrained-vit = "^0.0.7"
pandas = "^1.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
tensorboard = "^2.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
